removed the ability to cutomize cilium
The helm chart can generate certificate autorities which change on every
generation of the template.
This makes the generation of the machine configuration not
deterministic.
It also lead to problems if the configuration is reapplied with other
certificates.
Cilium is now bootstrapped as pure and minimal cni only.
Any modifications should happen Day 1.